refactor: review-comment: Fuse closely connected classes
