+++ src/opossum_lib/input_formats/opossum/services/opossum_format_reader.py
@@ -4,7 +4,7 @@
 from pathlib import Path
 
 from opossum_lib.core.entities.opossum import Opossum
-from opossum_lib.core.services.input_format_reader import InputFormatReader
+from opossum_lib.core.services.input_reader import InputFormatReader
 from opossum_lib.input_formats.opossum.services.opossum_file_reader import (
     OpossumFileReader,
 )
